unmodulize and link
