+++ client/components/comments.js
@@ -8,41 +8,23 @@ class Comments extends React.Component {
     super(props);
 
     this.state = {
-      comments: [],
-      body: '',
-      q1: '',
-      q2: ''
+      comment: ''
     };
 
     this.onChange = this.onChange.bind(this);
     this.onSubmit = this.onSubmit.bind(this);
   }
 
-  componentDidUpdate() {
-    if (
-      this.props.q1 !== undefined &&
-      this.props.q2 !== undefined
-      //   this.props.q1 !== this.state.q1) ||
-      // this.props.q2 !== this.state.q2
-    ) {
-      // this.setState({
-      //   comments: data,
-      //   q1: this.props.q1,
-      //   q2: this.props.q2
-      // });
-    }
-  }
-
   onChange(evt) {
-    this.setState({body: evt.target.value});
+    this.setState({comment: evt.target.value});
   }
 
   onSubmit(evt) {
     evt.preventDefault();
     console.log(this.props.q1, this.props.q2);
-    this.props.addComment(this.props.q1, this.props.q2, this.state.body);
+    this.props.addComment(this.props.q1, this.props.q2, this.state.comment);
     this.setState({
-      body: ''
+      comment: ''
     });
   }
 
@@ -56,7 +38,7 @@ class Comments extends React.Component {
                 <b>
                   {comment.user.email} ({comment.date}):{' '}
                 </b>
-                {comment.body}
+                {comment.comment}
               </p>
             ))}
           </div>
@@ -69,7 +51,7 @@ class Comments extends React.Component {
                       onChange={this.onChange}
                       id="textarea1"
                       className="materialize-textarea"
-                      value={this.state.body}
+                      value={this.state.comment}
                     ></textarea>
                     <label htmlFor="textarea1">Add a comment</label>
                   </div>
